Change address of user code to minimize search space
For 32-bit, the user code was mapped at 0x100000. This is a byte
combination that will be hit by the sifting algorithm and change the
exception. When this happens, we will classify this as "interesting"
and print this instruction.

By using an address that has more than one byte set, we can avoid this
scenario.